APPS-1284-astools-go - refactoreing - tests fix
