+++ cmd/internal/flags/compression.go
@@ -15,37 +15,32 @@
 package flags
 
 import (
-	"github.com/aerospike/backup-go/cmd/asbackup/models"
+	"github.com/aerospike/backup-go/cmd/internal/models"
 	"github.com/spf13/pflag"
 )
 
-type AwsS3 struct {
-	models.AwsS3
+type Compression struct {
+	models.Compression
 }
 
-func NewAwsS3() *AwsS3 {
-	return &AwsS3{}
+func NewCompression() *Compression {
+	return &Compression{}
 }
 
-func (f *AwsS3) NewFlagSet() *pflag.FlagSet {
+func (f *Compression) NewFlagSet() *pflag.FlagSet {
 	flagSet := &pflag.FlagSet{}
-
-	flagSet.StringVar(&f.Region, "s3-region",
-		"",
-		"The S3 region that the bucket(s) exist in.")
-	flagSet.StringVar(&f.Profile, "s3-profile",
-		"default",
-		"The S3 profile to use for credentials (the default is 'default').")
-	flagSet.StringVar(&f.Endpoint, "s3-endpoint-overrid",
-		"",
-		"An alternate url endpoint to send S3 API calls to.")
-	flagSet.IntVar(&f.MinPartSize, "s3-min-part-size",
-		3005,
-		"Secret agent port (only for TCP connection).")
+	flagSet.StringVarP(&f.Mode, "compress", "z",
+		"NONE",
+		"Enables compressing of backup files using the specified compression algorithm.\n"+
+			"Supported compression algorithms are: ZSTD, NONE\n"+
+			"Set the zstd compression level via the --compression-level option. Default level is 3.")
+	flagSet.IntVar(&f.Level, "compression-level",
+		3,
+		"zstd compression level.")
 
 	return flagSet
 }
 
-func (f *AwsS3) GetAwsS3() *models.AwsS3 {
-	return &f.AwsS3
+func (f *Compression) GetCompression() *models.Compression {
+	return &f.Compression
 }
