+++ cmd/internal/flags/compression_test.go
