+++ cmd/internal/flags/encryption_test.go
