+++ cmd/internal/flags/secret_agent_test.go
