+++ cmd/internal/models/compression.go
