+++ cmd/internal/models/encryption.go
