+++ cmd/internal/models/gcp_storage.go
