+++ tests/integration/s3_integration_test.go
@@ -3,10 +3,8 @@ package integration
 import (
 	"context"
 	"crypto/rand"
-	"encoding/json"
 	"fmt"
 	"io"
-	"log/slog"
 	"os"
 	"path/filepath"
 	"testing"
@@ -16,12 +14,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/docker/docker/api/types/container"
-	"github.com/docker/docker/api/types/image"
 	"github.com/docker/docker/client"
-	"github.com/docker/docker/pkg/jsonmessage"
-	"github.com/docker/go-connections/nat"
-	"github.com/minio/minio-go/v7"
-	"github.com/minio/minio-go/v7/pkg/credentials"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 )
@@ -44,76 +37,7 @@ func TestReadWrite(t *testing.T) {
 }
 
 func (s *writeReadTestSuite) SetupSuite() {
-	var err error
-	s.docker, err = client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
-	if err != nil {
-		s.FailNow("Failed to create Docker client", err)
-	}
-
-	ctx := context.Background()
-	responseBody, err := s.docker.ImagePull(ctx, "minio/minio", image.PullOptions{})
-	if err != nil {
-		s.FailNow("could not pull minio image", err)
-	}
-	defer func(responseBody io.ReadCloser) {
-		_ = responseBody.Close()
-	}(responseBody)
-
-	dec := json.NewDecoder(responseBody)
-	for {
-		var jm jsonmessage.JSONMessage
-		if err := dec.Decode(&jm); err != nil {
-			if err == io.EOF {
-				break
-			}
-			s.FailNow("could not decode pull image response", err)
-		}
-		slog.Info("Image pull", "status", jm)
-	}
-
-	minioResponse, err := s.docker.ContainerCreate(ctx, &container.Config{
-		Image: "minio/minio",
-		Cmd:   []string{"server", "/data", "--console-address", ":9001"},
-		Env: []string{
-			"MINIO_ROOT_USER=minioadmin",
-			"MINIO_ROOT_PASSWORD=minioadmin",
-		},
-		ExposedPorts: nat.PortSet{
-			"9001/tcp": struct{}{},
-		},
-	}, &container.HostConfig{
-		PortBindings: nat.PortMap{
-			"9000/tcp": []nat.PortBinding{
-				{HostIP: "0.0.0.0", HostPort: "9000"},
-			},
-			"9001/tcp": []nat.PortBinding{
-				{HostIP: "0.0.0.0", HostPort: "9001"},
-			},
-		},
-	}, nil, nil, "minio_test")
-	if err != nil {
-		s.FailNow("could not create minio container", err)
-	}
-
-	s.minioID = minioResponse.ID
-	if err := s.docker.ContainerStart(ctx, s.minioID, container.StartOptions{}); err != nil {
-		s.FailNow("could not start minio container", err)
-	}
-
-	minioClient, err := minio.New("localhost:9000", &minio.Options{
-		Creds: credentials.NewStaticV4("minioadmin", "minioadmin", ""),
-	})
-	if err != nil {
-		s.FailNow("could not create minio client", err)
-	}
-
-	err = minioClient.MakeBucket(ctx, "backup", minio.MakeBucketOptions{})
-	if err != nil {
-		s.FailNow("could not create bucket", err)
-	}
-
-	err = createMinioCredentialsFile()
-	if err != nil {
+	if err := createMinioCredentialsFile(); err != nil {
 		s.FailNow("could not create credentials file", err)
 	}
 }
@@ -135,7 +59,7 @@ func createMinioCredentialsFile() error {
 	if _, err := os.Stat(filePath); os.IsNotExist(err) {
 		credentialsFileBytes := []byte(`[minio]
 aws_access_key_id = minioadmin
-aws_secret_access_key = minioadmin`)
+aws_secret_access_key = minioadminpassword`)
 
 		err = os.WriteFile(filePath, credentialsFileBytes, 0o600)
 		if err != nil {
